[chore] moving webpack scripts to scripts folder.
