+++ scripts/webpack.dist.config.js
@@ -1,13 +1,14 @@
-var webpack = require('webpack');
-var path = require('path');
+const webpack = require('webpack');
+const path = require('path');
+const defaultConfig = require('./defaultConfig');
 
-var reactExternal = {
+const reactExternal = {
   root: 'React',
   commonjs2: 'react',
   commonjs: 'react',
   amd: 'react'
 };
-var reactDOMExternal = {
+const reactDOMExternal = {
   root: 'ReactDOM',
   commonjs2: 'react-dom',
   commonjs: 'react-dom',
@@ -15,41 +16,30 @@ var reactDOMExternal = {
 };
 
 module.exports = {
+  ...defaultConfig,
   mode: 'production',
-
   entry: {
-    'react-modal': './src/index.js',
-    'react-modal.min': './src/index.js'
+    'react-modal': path.resolve(__dirname, '../src/index.js'),
+    'react-modal.min': path.resolve(__dirname, '../src/index.js')
   },
-
   externals: {
     'react': reactExternal,
     'react-dom': reactDOMExternal
   },
-
   output: {
     filename: '[name].js',
     chunkFilename: '[id].chunk.js',
-    path: path.resolve(__dirname, 'dist'),
+    path: path.resolve(__dirname, '../dist'),
     publicPath: '/',
     libraryTarget: 'umd',
     library: 'ReactModal'
   },
-
   optimization: {
     minimize: true
   },
-
   plugins: [
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)
     })
-  ],
-
-  module: {
-    rules: [
-      { test: /\.js?$/, exclude: /node_modules/, use: { loader: 'babel-loader' } }
-    ]
-  }
-
+  ]
 };
