--- webpack.test.config.js
@@ -1,16 +1,0 @@
-const path = require('path');
-const commonConfig = require('./webpack.config');
-
-commonConfig.plugins = [];
-delete commonConfig.optimization;
-commonConfig.entry = './specs/index.js';
-commonConfig.devtool = 'inline-source-map';
-
-commonConfig.module.rules.unshift({
-  test: /\.js$/,
-  use: { loader: 'istanbul-instrumenter-loader' },
-  enforce: 'post',
-  include: path.resolve(__dirname, './src')
-});
-
-module.exports = commonConfig;
